Refactor Bindings class into individual mode classes
Whereas before the Bindings class was one huge class implementing all the
possible modes and their support methods, this PR breaks up the Bindings
class into several small modules (under ginga.modes) each one of which
implements a particular mode.  It also standardizes a class structure for
implementing a mode, making it somewhat similar to a plugin (start and
stop methods, etc).  As a result, it is much easier for a developer to
understand how to modify a mode or to make a new one.

The mode indicator is also modified to become a mode line, so that the
on screen display of adjustments can be more easily read (the previous
readout was often difficult to read depending on the brightness or color
of the background).

Modes are a feature of the Ginga widget and can be used wherever the widget
is used so this PR affects all users not just reference viewer users.

This explicitly breaks the API for the ginga.Bindings module and that is
intentional.  From the user's perspective of using a "bindings.cfg" file
however, nothing has changed.
