+++ ginga/modes/modeinfo.py
@@ -1,0 +1,12 @@
+from ginga.modes.cuts import CutsMode
+from ginga.modes.contrast import ContrastMode
+from ginga.modes.rotate import RotateMode
+from ginga.modes.cmap import CMapMode
+from ginga.modes.dist import DistMode
+from ginga.modes.camera import CameraMode
+from ginga.modes.naxis import NaxisMode
+from ginga.modes.pan import PanMode
+from ginga.modes.freepan import FreePanMode
+
+available_modes = [CutsMode, ContrastMode, RotateMode, CMapMode,
+                   DistMode, CameraMode, NaxisMode, PanMode, FreePanMode]
