+++ ginga/qtw/ImageViewQt.py
@@ -983,8 +983,7 @@ class ScrolledView(QtGui.QAbstractScrollArea):
         # flag that suppresses a cyclical callback
         self._adjusting = True
         try:
-            bd = self.viewer.get_bindings()
-            res = bd.calc_pan_pct(self.viewer, pad=self.pad)
+            res = self.viewer.calc_pan_pct(pad=self.pad)
             if res is None:
                 return
 
@@ -1019,8 +1018,7 @@ class ScrolledView(QtGui.QAbstractScrollArea):
 
         self._scrolling = True
         try:
-            bd = self.viewer.get_bindings()
-            res = bd.calc_pan_pct(self.viewer, pad=self.pad)
+            res = self.viewer.calc_pan_pct(pad=self.pad)
             if res is None:
                 return
             pct_x, pct_y = res.pan_pct_x, res.pan_pct_y
@@ -1036,8 +1034,7 @@ class ScrolledView(QtGui.QAbstractScrollArea):
                 # invert Y pct because of orientation of scrollbar
                 pct_y = 1.0 - (pos_y / float(self.upper_v))
 
-            bd = self.viewer.get_bindings()
-            bd.pan_by_pct(self.viewer, pct_x, pct_y, pad=self.pad)
+            self.viewer.pan_by_pct(pct_x, pct_y, pad=self.pad)
 
             # This shouldn't be necessary, but seems to be
             self.viewer.redraw(whence=0)
